requirements.txt
flask>=3.0.0
anthropic>=0.40.0
python-docx>=1.1.0
mammoth>=1.8.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0
python-dotenv>=1.0.0
